Refactor tenancy packages
