+++ cmd/collector/app/span_processor.go
@@ -27,7 +27,7 @@ import (
 	"github.com/jaegertracing/jaeger/cmd/collector/app/sanitizer"
 	"github.com/jaegertracing/jaeger/model"
 	"github.com/jaegertracing/jaeger/pkg/queue"
-	"github.com/jaegertracing/jaeger/storage"
+	"github.com/jaegertracing/jaeger/pkg/tenancy"
 	"github.com/jaegertracing/jaeger/storage/spanstore"
 )
 
@@ -156,7 +156,7 @@ func (sp *spanProcessor) saveSpan(span *model.Span, tenant string) {
 	// Since we save spans asynchronously from receiving them, we cannot reuse
 	// the inbound Context, as it may be cancelled by the time we reach this point,
 	// so we need to start a new Context.
-	ctx := storage.WithTenant(context.Background(), tenant)
+	ctx := tenancy.WithTenant(context.Background(), tenant)
 	if err := sp.spanWriter.WriteSpan(ctx, span); err != nil {
 		sp.logger.Error("Failed to save span", zap.Error(err))
 		sp.metrics.SavedErrBySvc.ReportServiceNameForSpan(span)
